+++ lib/accounts.dart
@@ -10,6 +10,7 @@ import 'enums.dart';
 import 'http_client_wrapper.dart';
 import 'http_element_parser.dart';
 import 'global_url.dart';
+import 'lib_exception.dart';
 import 'range_class.dart';
 import 'subject_code.dart';
 
@@ -53,7 +54,8 @@ class Accounts {
     return AccountSession(
       sessionId: _getSessionIdFromHeader(response.setHeaders),
       viewState: docHtml.getElementById("__VIEWSTATE").getValue(),
-      viewStateGenerator: docHtml.getElementById("__VIEWSTATEGENERATOR").getValue(),
+      viewStateGenerator:
+          docHtml.getElementById("__VIEWSTATEGENERATOR").getValue(),
     );
   }
 
@@ -64,7 +66,9 @@ class Accounts {
     session.ensureValidSessionId();
 
     // Header data
-    Map<String, String> headers = <String, String>{'cookie': _sessionToCookieItem(session.sessionId!)};
+    Map<String, String> headers = <String, String>{
+      'cookie': _sessionToCookieItem(session.sessionId!)
+    };
 
     final response = await HttpClientWrapper.get(
       uri: Uri.parse(GlobalUrl.subjectScheduleLink(year: 22, semester: 1)),
@@ -74,7 +78,8 @@ class Accounts {
 
     if (response.ex != null) {
       return LoginStatus.unknown;
-    } else if ((response.statusCode ?? 0) >= 200 && (response.statusCode ?? 0) <= 299) {
+    } else if ((response.statusCode ?? 0) >= 200 &&
+        (response.statusCode ?? 0) <= 299) {
       return LoginStatus.loggedIn;
     } else {
       return LoginStatus.loggedOut;
@@ -119,7 +124,9 @@ class Accounts {
     session.ensureValidSessionId();
 
     // Header data
-    Map<String, String> headers = <String, String>{'cookie': _sessionToCookieItem(session.sessionId!)};
+    Map<String, String> headers = <String, String>{
+      'cookie': _sessionToCookieItem(session.sessionId!)
+    };
 
     await HttpClientWrapper.get(
       uri: Uri.parse(GlobalUrl.logoutLink()),
@@ -137,14 +144,20 @@ class Accounts {
     session.ensureValidSessionId();
 
     if (await Accounts.isLoggedIn(session: session) != LoginStatus.loggedIn) {
-      throw Exception('You\'re not logged in!');
+      throw DutWrapperException(
+        message: "You're not logged in! Please login first.",
+        reason: DutWrapperExceptionReason.notAuthorized,
+      );
     }
 
     // Header data
-    Map<String, String> headers = <String, String>{'cookie': _sessionToCookieItem(session.sessionId!)};
+    Map<String, String> headers = <String, String>{
+      'cookie': _sessionToCookieItem(session.sessionId!)
+    };
 
     final response = await HttpClientWrapper.get(
-      uri: Uri.parse(GlobalUrl.subjectScheduleLink(year: year, semester: semester)),
+      uri: Uri.parse(
+          GlobalUrl.subjectScheduleLink(year: year, semester: semester)),
       headers: headers,
     ).timeout(Duration(seconds: timeout));
     response.ensureSuccessfulStatusCode();
@@ -183,7 +196,8 @@ class Accounts {
           // Will be replaced with following regex after spliited by "; "
           // (Thứ [2-7]|CN),([0-9]{1,2}-[0-9]{1,2}),(.*)
           if (schCell[7].text.isNotEmpty) {
-            RegExp regex = RegExp("(Thứ [2-7]|CN),([0-9]{1,2})-([0-9]{1,2}),(.*)");
+            RegExp regex =
+                RegExp("(Thứ [2-7]|CN),([0-9]{1,2})-([0-9]{1,2}),(.*)");
             schCell[7].text.split('; ').forEach((element) {
               if (regex.hasMatch(element)) {
                 item.subjectStudy.subjectStudyList.add(
@@ -192,10 +206,17 @@ class Accounts {
                         ? 1
                         : regex.firstMatch(element)?.group(1) == "CN"
                             ? 1
-                            : int.parse(regex.firstMatch(element)!.group(1)!.split(' ')[1]),
+                            : int.parse(regex
+                                .firstMatch(element)!
+                                .group(1)!
+                                .split(' ')[1]),
                     lesson: RangeInt(
-                      start: int.tryParse(regex.firstMatch(element)?.group(2) ?? "") ?? 0,
-                      end: int.tryParse(regex.firstMatch(element)?.group(3) ?? "") ?? 0,
+                      start: int.tryParse(
+                              regex.firstMatch(element)?.group(2) ?? "") ??
+                          0,
+                      end: int.tryParse(
+                              regex.firstMatch(element)?.group(3) ?? "") ??
+                          0,
                     ),
                     room: regex.firstMatch(element)?.group(4) ?? "",
                   ),
@@ -232,27 +253,32 @@ class Accounts {
           }
 
           try {
-            SubjectInformation schItem =
-                result.firstWhere((element) => element.id == SubjectCode.fromString(input: schCell[1].text));
+            SubjectInformation schItem = result.firstWhere((element) =>
+                element.id == SubjectCode.fromString(input: schCell[1].text));
 
             // Will be replaced with following regex:
             // Ngày: ([0-9]{2}\/[0-9]{2}\/[0-9]{4}), Phòng: (.*), Giờ: ([0-9]{1,2}h[0-9]{2}), Xuất: (.*)
 
             if (schCell[5].text.isNotEmpty) {
-              RegExp regex =
-                  RegExp(r"Ngày: ([0-9]{2}/[0-9]{2}/[0-9]{4}), Phòng: (.*), Giờ: ([0-9]{1,2}h[0-9]{2}), Xuất: (.*)");
+              RegExp regex = RegExp(
+                  r"Ngày: ([0-9]{2}/[0-9]{2}/[0-9]{4}), Phòng: (.*), Giờ: ([0-9]{1,2}h[0-9]{2}), Xuất: (.*)");
               if (regex.hasMatch(schCell[5].text)) {
-                var dateSplit = regex.firstMatch(schCell[5].text)!.group(1)!.split("/");
+                var dateSplit =
+                    regex.firstMatch(schCell[5].text)!.group(1)!.split("/");
                 var timeSplit = regex.firstMatch(schCell[5].text)!.group(3)!;
 
-                DateTime dt = DateTime.parse("${dateSplit[2]}-${dateSplit[1]}-${dateSplit[0]}");
+                DateTime dt = DateTime.parse(
+                    "${dateSplit[2]}-${dateSplit[1]}-${dateSplit[0]}");
                 dt = dt.add(Duration(
                   hours: int.parse(timeSplit.split("h")[0]),
-                  minutes: timeSplit.split("h").length == 2 ? int.parse(timeSplit.split("h")[1]) : 0,
+                  minutes: timeSplit.split("h").length == 2
+                      ? int.parse(timeSplit.split("h")[1])
+                      : 0,
                 ));
 
                 schItem.subjectExam.date = dt.millisecondsSinceEpoch;
-                schItem.subjectExam.room = regex.firstMatch(schCell[5].text)?.group(2) ?? "";
+                schItem.subjectExam.room =
+                    regex.firstMatch(schCell[5].text)?.group(2) ?? "";
                 // Is global - Get from root element
                 schItem.subjectExam.isGlobal = schCell[4].isGridChecked();
                 // Group - Get from root element
@@ -280,11 +306,16 @@ class Accounts {
     session.ensureValidSessionId();
 
     if (await Accounts.isLoggedIn(session: session) != LoginStatus.loggedIn) {
-      throw Exception('You\'re not logged in!');
+      throw DutWrapperException(
+        message: "You're not logged in! Please login first.",
+        reason: DutWrapperExceptionReason.notAuthorized,
+      );
     }
 
     // Header data
-    Map<String, String> headers = <String, String>{'cookie': _sessionToCookieItem(session.sessionId!)};
+    Map<String, String> headers = <String, String>{
+      'cookie': _sessionToCookieItem(session.sessionId!)
+    };
 
     final response = await HttpClientWrapper.get(
       uri: Uri.parse(GlobalUrl.subjectFeeLink(year: year, semester: semester)),
@@ -348,11 +379,16 @@ class Accounts {
     session.ensureValidSessionId();
 
     if (await Accounts.isLoggedIn(session: session) != LoginStatus.loggedIn) {
-      throw Exception('You\'re not logged in!');
+      throw DutWrapperException(
+        message: "You're not logged in! Please login first.",
+        reason: DutWrapperExceptionReason.notAuthorized,
+      );
     }
 
     // Header data
-    Map<String, String> headers = <String, String>{'cookie': _sessionToCookieItem(session.sessionId!)};
+    Map<String, String> headers = <String, String>{
+      'cookie': _sessionToCookieItem(session.sessionId!)
+    };
 
     final response = await HttpClientWrapper.get(
       uri: Uri.parse(GlobalUrl.accountInformationLink()),
@@ -362,7 +398,12 @@ class Accounts {
 
     // Check if response have content here. If not, just throw errors.
     if (response.body == null) {
-      throw Exception('We can\'t receive any information about this request! Please, try again.');
+      throw DutWrapperException(
+        message: "We can't receive any information about this request! "
+            "Please, try again later.\n"
+            "You might need to check if you're logged in first.",
+        reason: DutWrapperExceptionReason.dataNotFoundException,
+      );
     }
 
     // Main processing
@@ -370,35 +411,69 @@ class Accounts {
     return StudentInformation(
       name: webDoc.getElementById("CN_txtHoTen").getValueOrEmpty(),
       dateOfBirth: webDoc.getElementById("CN_txtNgaySinh").getValueOrEmpty(),
-      birthPlace: webDoc.getElementById("CN_cboNoiSinh").getSelectedOptionInComboBox().getTextOrEmpty(),
+      birthPlace: webDoc
+          .getElementById("CN_cboNoiSinh")
+          .getSelectedOptionInComboBox()
+          .getTextOrEmpty(),
       gender: webDoc.getElementById("CN_txtGioiTinh").getValueOrEmpty(),
-      ethnicity: webDoc.getElementById("CN_cboDanToc").getSelectedOptionInComboBox().getTextOrEmpty(),
-      nationality: webDoc.getElementById("CN_cboQuocTich").getSelectedOptionInComboBox().getTextOrEmpty(),
+      ethnicity: webDoc
+          .getElementById("CN_cboDanToc")
+          .getSelectedOptionInComboBox()
+          .getTextOrEmpty(),
+      nationality: webDoc
+          .getElementById("CN_cboQuocTich")
+          .getSelectedOptionInComboBox()
+          .getTextOrEmpty(),
       nationalIdCard: webDoc.getElementById("CN_txtSoCMND").getValueOrEmpty(),
-      nationalIdCardIssueDate: webDoc.getElementById("CN_txtNgayCap").getValueOrEmpty(),
-      nationalIdCardIssuePlace: webDoc.getElementById("CN_cboNoiCap").getSelectedOptionInComboBox().getTextOrEmpty(),
+      nationalIdCardIssueDate:
+          webDoc.getElementById("CN_txtNgayCap").getValueOrEmpty(),
+      nationalIdCardIssuePlace: webDoc
+          .getElementById("CN_cboNoiCap")
+          .getSelectedOptionInComboBox()
+          .getTextOrEmpty(),
       citizenIdCard: webDoc.getElementById("CN_txtSoCCCD").getValueOrEmpty(),
-      citizenIdCardIssueDate: webDoc.getElementById("CN_txtNcCCCD").getValueOrEmpty(),
-      religion: webDoc.getElementById("CN_cboTonGiao").getSelectedOptionInComboBox().getTextOrEmpty(),
+      citizenIdCardIssueDate:
+          webDoc.getElementById("CN_txtNcCCCD").getValueOrEmpty(),
+      religion: webDoc
+          .getElementById("CN_cboTonGiao")
+          .getSelectedOptionInComboBox()
+          .getTextOrEmpty(),
       accountBankId: webDoc.getElementById("CN_txtTKNHang").getValueOrEmpty(),
       accountBankName: webDoc.getElementById("CN_txtNgHang").getValueOrEmpty(),
       hIId: webDoc.getElementById("CN_txtSoBHYT").getValueOrEmpty(),
       hIExpireDate: webDoc.getElementById("CN_txtHanBHYT").getValueOrEmpty(),
-      specialization: webDoc.getElementById("MainContent_CN_txtNganh").getValueOrEmpty(),
+      specialization:
+          webDoc.getElementById("MainContent_CN_txtNganh").getValueOrEmpty(),
       schoolClass: webDoc.getElementById("CN_txtLop").getValueOrEmpty(),
-      trainingProgramPlan: webDoc.getElementById("MainContent_CN_txtCTDT").getValueOrEmpty(),
-      trainingProgramPlan2: webDoc.getElementById("MainContent_CN_txtCT2").getValueOrEmpty(),
+      trainingProgramPlan:
+          webDoc.getElementById("MainContent_CN_txtCTDT").getValueOrEmpty(),
+      trainingProgramPlan2:
+          webDoc.getElementById("MainContent_CN_txtCT2").getValueOrEmpty(),
       schoolEmail: webDoc.getElementById("CN_txtMail1").getValueOrEmpty(),
       personalEmail: webDoc.getElementById("CN_txtMail2").getValueOrEmpty(),
-      schoolEmailInitPass: webDoc.getElementById("CN_txtMK365").getValueOrEmpty(),
+      schoolEmailInitPass:
+          webDoc.getElementById("CN_txtMK365").getValueOrEmpty(),
       facebookUrl: webDoc.getElementById("CN_txtFace").getValueOrEmpty(),
       phoneNumber: webDoc.getElementById("CN_txtPhone").getValueOrEmpty(),
       address: webDoc.getElementById("CN_txtCuTru").getValueOrEmpty(),
-      addressFrom: webDoc.getElementById("CN_cboDCCua").getSelectedOptionInComboBox().getTextOrEmpty(),
-      addressCity: webDoc.getElementById("CN_cboTinhCTru").getSelectedOptionInComboBox().getTextOrEmpty(),
-      addressDistrict: webDoc.getElementById("CN_cboQuanCTru").getSelectedOptionInComboBox().getTextOrEmpty(),
-      addressSubDistrict: webDoc.getElementById("CN_divPhuongCTru").getSelectedOptionInComboBox().getTextOrEmpty(),
-      studentId: _parseStudentId(webDoc.getElementById("Main_lblHoTen").getTextOrEmpty()),
+      addressFrom: webDoc
+          .getElementById("CN_cboDCCua")
+          .getSelectedOptionInComboBox()
+          .getTextOrEmpty(),
+      addressCity: webDoc
+          .getElementById("CN_cboTinhCTru")
+          .getSelectedOptionInComboBox()
+          .getTextOrEmpty(),
+      addressDistrict: webDoc
+          .getElementById("CN_cboQuanCTru")
+          .getSelectedOptionInComboBox()
+          .getTextOrEmpty(),
+      addressSubDistrict: webDoc
+          .getElementById("CN_divPhuongCTru")
+          .getSelectedOptionInComboBox()
+          .getTextOrEmpty(),
+      studentId: _parseStudentId(
+          webDoc.getElementById("Main_lblHoTen").getTextOrEmpty()),
     );
   }
 
@@ -409,11 +484,16 @@ class Accounts {
     session.ensureValidSessionId();
 
     if (await Accounts.isLoggedIn(session: session) != LoginStatus.loggedIn) {
-      throw Exception('You\'re not logged in!');
+      throw DutWrapperException(
+        message: "You're not logged in! Please login first.",
+        reason: DutWrapperExceptionReason.notAuthorized,
+      );
     }
 
     // Header data
-    Map<String, String> headers = <String, String>{'cookie': _sessionToCookieItem(session.sessionId!)};
+    Map<String, String> headers = <String, String>{
+      'cookie': _sessionToCookieItem(session.sessionId!)
+    };
 
     final response = await HttpClientWrapper.get(
       uri: Uri.parse(GlobalUrl.trainingStatusLink()),
@@ -424,7 +504,12 @@ class Accounts {
     // Check if response have content here.
     // If not, just throw errors.
     if (response.body == null) {
-      throw Exception('We can\'t receive any information about this request! Please, try again.');
+      throw DutWrapperException(
+        message: "We can't receive any information about this request! "
+            "Please, try again later.\n"
+            "You might need to check if you're logged in first.",
+        reason: DutWrapperExceptionReason.dataNotFoundException,
+      );
     }
 
     // Main processing
@@ -449,9 +534,13 @@ class Accounts {
       avgTrainingScore4: 0,
       avgSocial: 0,
     );
-    final webDocTrainSum = webDoc.getElementById("KQRLGridTH")?.getElementsByClassName("GridRow");
+    final webDocTrainSum =
+        webDoc.getElementById("KQRLGridTH")?.getElementsByClassName("GridRow");
     if (webDocTrainSum == null) {
-      throw Exception("No data for training summary");
+      throw DutWrapperException(
+        message: "No data for training summary.",
+        reason: DutWrapperExceptionReason.dataNotFoundException,
+      );
     }
     for (var gridRow in webDocTrainSum) {
       var gridCell = gridRow.getElementsByClassName("GridCell");
@@ -463,11 +552,19 @@ class Accounts {
       }
 
       trainSum = trainSum.copyWith(
-        schoolYearStart: (trainSum.schoolYearStart.isEmpty) ? gridCell[0].getTextOrEmpty() : trainSum.schoolYearStart,
+        schoolYearStart: (trainSum.schoolYearStart.isEmpty)
+            ? gridCell[0].getTextOrEmpty()
+            : trainSum.schoolYearStart,
         schoolYearCurrent: gridCell[0].getTextOrEmpty(),
-        creditCollected: double.tryParse(gridCell[gridCell.length - 3].getTextOrEmpty()) ?? 0,
-        avgTrainingScore4: double.tryParse(gridCell[gridCell.length - 2].getTextOrEmpty()) ?? 0,
-        avgSocial: double.tryParse(gridCell[gridCell.length - 1].getTextOrEmpty()) ?? 0,
+        creditCollected:
+            double.tryParse(gridCell[gridCell.length - 3].getTextOrEmpty()) ??
+                0,
+        avgTrainingScore4:
+            double.tryParse(gridCell[gridCell.length - 2].getTextOrEmpty()) ??
+                0,
+        avgSocial:
+            double.tryParse(gridCell[gridCell.length - 1].getTextOrEmpty()) ??
+                0,
       );
     }
 
@@ -476,25 +573,44 @@ class Accounts {
 
   static GraduateStatus _parseGraduateStatus({required Document webDoc}) {
     // Gradudate status
-    var webDocGraduateStat = webDoc.getElementById("KQRLdvCc").convertToDocument();
+    var webDocGraduateStat =
+        webDoc.getElementById("KQRLdvCc").convertToDocument();
     if (webDocGraduateStat == null) {
-      throw Exception("No data for gradudate status");
+      throw DutWrapperException(
+        message: "No data for gradudate status.",
+        reason: DutWrapperExceptionReason.dataNotFoundException,
+      );
     }
     GraduateStatus gradStat = GraduateStatus(
       hasSigGDTC: webDocGraduateStat.getElementById("KQRL_chkGDTC").isChecked(),
       hasSigGDQP: webDocGraduateStat.getElementById("KQRL_chkQP").isChecked(),
-      hasSigEnglish: webDocGraduateStat.getElementById("KQRL_chkCCNN").isChecked(),
+      hasSigEnglish:
+          webDocGraduateStat.getElementById("KQRL_chkCCNN").isChecked(),
       hasSigIT: webDocGraduateStat.getElementById("KQRL_chkCCTH").isChecked(),
-      hasQualifiedGraduate: webDocGraduateStat.getElementById("KQRL_chkCNTN").isChecked(),
-      rewardsInfo: webDocGraduateStat.getElementById("KQRL_txtKT").getTextOrEmpty().trim(),
-      disciplineInfo: webDocGraduateStat.getElementById("KQRL_txtKL").getTextOrEmpty().trim(),
-      eligibleGraduationThesisStatus: webDocGraduateStat.getElementById("KQRL_txtInfo").getTextOrEmpty().trim(),
-      eligibleGraduationStatus: webDocGraduateStat.getElementById("KQRL_txtCNTN").getTextOrEmpty().trim(),
+      hasQualifiedGraduate:
+          webDocGraduateStat.getElementById("KQRL_chkCNTN").isChecked(),
+      rewardsInfo: webDocGraduateStat
+          .getElementById("KQRL_txtKT")
+          .getTextOrEmpty()
+          .trim(),
+      disciplineInfo: webDocGraduateStat
+          .getElementById("KQRL_txtKL")
+          .getTextOrEmpty()
+          .trim(),
+      eligibleGraduationThesisStatus: webDocGraduateStat
+          .getElementById("KQRL_txtInfo")
+          .getTextOrEmpty()
+          .trim(),
+      eligibleGraduationStatus: webDocGraduateStat
+          .getElementById("KQRL_txtCNTN")
+          .getTextOrEmpty()
+          .trim(),
     );
     return gradStat;
   }
 
-  static List<SubjectResult> _parseSubjectResultList({required Document webDoc}) {
+  static List<SubjectResult> _parseSubjectResultList(
+      {required Document webDoc}) {
     // Subject result list
     var webDocSubjectResult = webDoc
         .getElementById("KQRL_divContent")
@@ -502,7 +618,10 @@ class Accounts {
         ?.getElementById("KQRLGridKQHT")
         ?.getElementsByClassName("GridRow");
     if (webDocSubjectResult == null) {
-      throw Exception("No data for subject result list");
+      throw DutWrapperException(
+        message: "No data for subject result list.",
+        reason: DutWrapperExceptionReason.dataNotFoundException,
+      );
     }
     List<SubjectResult> subjectResultList = [];
     for (var row in webDocSubjectResult.reversed) {
@@ -530,8 +649,9 @@ class Accounts {
         resultT10: double.tryParse(gridCell[15].getTextOrEmpty()),
         resultT4: double.tryParse(gridCell[16].getTextOrEmpty()),
         resultByCharacter: gridCell[17].getText(),
-        isReStudy: subjectResultList.firstWhereOrNull(
-                (element) => element.name.toLowerCase() == gridCell[4].getTextOrEmpty().toLowerCase()) !=
+        isReStudy: subjectResultList.firstWhereOrNull((element) =>
+                element.name.toLowerCase() ==
+                gridCell[4].getTextOrEmpty().toLowerCase()) !=
             null,
       ));
     }