+++ lib/utils.dart
@@ -1,9 +1,13 @@
+import 'dart:io';
+
+import 'package:flutter/foundation.dart';
 import 'package:html/parser.dart';
 import 'package:http/http.dart' as http;
 
 import 'http_client_wrapper.dart';
 import 'http_element_parser.dart';
 import 'global_url.dart';
+import 'lib_exception.dart';
 import 'utils_object.dart';
 
 class Utils {
@@ -17,7 +21,9 @@ class Utils {
 
   static Future<DutSchoolYear?> getCurrentSchoolYear({int timeout = 60}) async {
     try {
-      final response = await http.get(Uri.parse(GlobalUrl.dutSchedulePage())).timeout(Duration(seconds: timeout));
+      final response = await http
+          .get(Uri.parse(GlobalUrl.dutSchedulePage()))
+          .timeout(Duration(seconds: timeout));
 
       // Main processing
       var webDoc = parse(response.body);
@@ -28,22 +34,33 @@ class Utils {
       int? week;
 
       // School year item processing
-      var cbbYear = webDoc.getElementById("dnn_ctr442_View_cboNamhoc").getSelectedOptionInComboBox();
+      var cbbYear = webDoc
+          .getElementById("dnn_ctr442_View_cboNamhoc")
+          .getSelectedOptionInComboBox();
       if (cbbYear == null) {
-        // TODO: Error while parsing here.
-        throw Exception("");
+        throw DutWrapperException(
+          message: "We can't receive any information about this request! "
+              "Please, try again later.",
+          reason: DutWrapperExceptionReason.dataNotFoundException,
+        );
       } else {
         schYear = cbbYear.getText();
         schYearVal = int.parse(cbbYear.getValue() ?? "0");
       }
 
       // Week item processing
-      var cbbWeek = webDoc.getElementById("dnn_ctr442_View_cboTuan").getSelectedOptionInComboBox();
+      var cbbWeek = webDoc
+          .getElementById("dnn_ctr442_View_cboTuan")
+          .getSelectedOptionInComboBox();
       if (cbbWeek == null) {
-        // TODO: Error while parsing here.
-        throw Exception("");
+        throw DutWrapperException(
+          message: "We can't receive any information about this request! "
+              "Please, try again later.",
+          reason: DutWrapperExceptionReason.dataNotFoundException,
+        );
       } else {
-        RegExp regex = RegExp("Tuần thứ (\\d{1,2}): (\\d{1,2}\\/\\d{1,2}\\/\\d{4})");
+        RegExp regex =
+            RegExp("Tuần thứ (\\d{1,2}): (\\d{1,2}\\/\\d{1,2}\\/\\d{4})");
         if (regex.hasMatch(cbbWeek.text)) {
           var match1 = regex.firstMatch(cbbWeek.text)!;
           week = int.parse(match1.group(1)!);
@@ -71,7 +88,15 @@ class Utils {
     int dayOfweek = 1,
     bool fullString = false,
   }) {
-    var dataFull = ["Sunday", "Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday"];
+    var dataFull = [
+      "Sunday",
+      "Monday",
+      "Tuesday",
+      "Wednesday",
+      "Thursday",
+      "Friday",
+      "Saturday"
+    ];
     var dataShort = ["Sun", "Mon", "Tue", "Wed", "Thu", "Fri", "Sat"];
 
     if (dayOfweek > 7 || dayOfweek < 1) {
@@ -79,5 +104,39 @@ class Utils {
     }
 
     return fullString ? dataFull[dayOfweek - 1] : dataShort[dayOfweek - 1];
+  }
+
+  static Future<bool> _checkDomain(String domain) async {
+    bool isConnected = false;
+    try {
+      final result = await InternetAddress.lookup(domain);
+      if (result.isNotEmpty && result[0].rawAddress.isNotEmpty) {
+        isConnected = true;
+      }
+    } on SocketException catch (_) {
+      isConnected = false;
+    }
+    return isConnected;
+  }
+
+  static Future<void> ensureServerWorking() async {
+    debugPrint("Checking internet access...");
+    if (!await _checkDomain("example.com")) {
+      throw DutWrapperException(
+        message: "Looks like you don't have an internet connection. "
+            "Check your internet settings, and try again.",
+        reason: DutWrapperExceptionReason.internetNotFound,
+      );
+    }
+    debugPrint("Checking sv.dut.udn.vn...");
+    if (!await _checkDomain("sv.dut.udn.vn")) {
+      throw DutWrapperException(
+        message: "Looks like you have an internet connection, "
+            "but can't connect to sv.dut.udn.vn server. Try again later. "
+            "You might need to check your internet settings again to confirm.",
+        reason: DutWrapperExceptionReason.internetNotFound,
+      );
+    }
+    debugPrint("It looks like you have connected to sv.dut.udn.vn.");
   }
 }